Files separation into specifics modules
