--- modules/algorithms/include/algorithms/oaa_classifier.hpp
@@ -1,87 +1,0 @@
-/*L*************************************************************************************************
-*
-*  IMPORTANT: READ BEFORE DOWNLOADING, COPYING, INSTALLING OR USING.
-* 
-*  By downloading, copying, installing or using the software you agree to this license. If you do 
-*  not agree to this license, do not download, install, copy or use the software.
-*
-*                            Software License Agreement (BSD License)
-*                               For Smart Surveillance Framework
-*                                 http://ssig.dcc.ufmg.br/ssf/
-*
-*  Copyright (c) 2013, Smart Surveillance Interest Group, all rights reserved.
-*  
-*  Redistribution and use in source and binary forms, with or without modification, are permitted 
-*  provided that the following conditions are met:
-*
-*    1. Redistributions of source code must retain the above copyright notice, this list of 
-*       conditions and the following disclaimer.
-*
-*    2. Redistributions in binary form must reproduce the above copyright notice, this list of 
-*       conditions and the following disclaimer in the documentation and/or other materials 
-*       provided with the distribution.
-*
-*    3. Neither the name of the copyright holder nor the names of its contributors may be used to 
-*       endorse or promote products derived from this software without specific prior written 
-*       permission.
-*
-*  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR 
-*  IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY 
-*  AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
-*  CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR 
-*  CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR 
-*  SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY 
-*  THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR 
-*  OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE 
-*  POSSIBILITY OF SUCH DAMAGE.
-*************************************************************************************************L*/
-
-#ifndef _SSF_ALGORITHMS_OAACLASSIFIER_HPP_
-#define _SSF_ALGORITHMS_OAACLASSIFIER_HPP_
-#include "classification.hpp"
-#include <unordered_map>
-#include <memory>
-
-namespace ssf{
-
-
-class OAAClassifier : public Classification{
-  ALG_EXPORT virtual void addLabels(cv::Mat_<int>& labels);
-public:
-  ALG_EXPORT OAAClassifier(const Classification& prototypeClassifier);
-  virtual ~OAAClassifier(void) = default;
-
-  ALG_EXPORT virtual void predict(cv::Mat_<float>& inp, cv::Mat_<float>& resp) const override;
-
-  ALG_EXPORT virtual void learn(cv::Mat_<float>& input, cv::Mat_<int>& labels) override;
-  ALG_EXPORT virtual cv::Mat_<int> getLabels() const override;
-  ALG_EXPORT virtual std::unordered_map<int, int> getLabelsOrdering() const override;
-  ALG_EXPORT virtual bool empty() const override;
-  ALG_EXPORT virtual bool isTrained() const override;
-  ALG_EXPORT virtual bool isClassifier() const override;
-  ALG_EXPORT virtual void load(const std::string& filename, const std::string& nodename) override;
-  ALG_EXPORT virtual void save(const std::string& filename, const std::string& nodename) const override;
-
-  ALG_EXPORT virtual void read(const cv::FileNode& fn) override;
-  ALG_EXPORT virtual void write(cv::FileStorage& fs) const override;
-
-  ALG_EXPORT virtual Classification* clone() const override;
-
-  ALG_EXPORT std::shared_ptr<Classification> getUnderlyingClassifier() const;
-  ALG_EXPORT void setUnderlyingClassifier(const Classification& underlyingClassifier);
-protected:
-  OAAClassifier() = default;
-
-private:
-  //private members
-  std::unordered_map<int, int> mLabelOrderings;
-  std::vector<std::shared_ptr<Classification>> mClassifiers;
-  std::unique_ptr<Classification> mUnderlyingClassifier;
-  cv::Mat_<int> mLabels;
-
-  bool mTrained = false;
-};
-
-}
-#endif // !_SSF_ALGORITHMS_OAACLASSIFIER_HPP_
-
